third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCPeerConnection-setRemoteDescription-replaceTrack.https-expected.txt

This is a testharness.js-based test.
PASS replaceTrack() sets the track attribute to a new track.
PASS replaceTrack() sets the track attribute to null.
PASS replaceTrack() does not set the track synchronously.
PASS replaceTrack() rejects when the peer connection is closed.
PASS replaceTrack() rejects when invoked after removeTrack().
FAIL replaceTrack() rejects after a subsequent removeTrack(). assert_unreached: Expected replaceTrack() to be rejected with InvalidModificationError but the promise was resolved. Reached unreachable code
Harness: the test ran to completion.
